[PERPICK-045] (statistic) StarRatingStatisticDomainService 제거
