+++ src/main/java/com/pikachu/purple/application/perfume/service/application/perfume/RecalculatePerfumeAverageScoreService.java
@@ -21,7 +21,7 @@ class RecalculatePerfumeAverageScoreService implements
     @Transactional
     public void invoke(Long perfumeId) {
         List<StarRatingStatistic> starRatingStatistics = getStarRatingStatisticsUseCase
-            .invoke(perfumeId)
+            .findAll(perfumeId)
             .starRatingStatistics();
 
         double totalScore = starRatingStatistics.stream()
