--- src/main/java/com/pikachu/purple/application/statistic/service/application/starratingstatistic/IncreaseStarRatingStatisticService.java
@@ -1,27 +1,0 @@
-package com.pikachu.purple.application.statistic.service.application.starratingstatistic;
-
-import com.pikachu.purple.application.statistic.port.in.starratingstatistic.IncreaseStarRatingStatisticUseCase;
-import com.pikachu.purple.application.statistic.service.domain.StarRatingStatisticDomainService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-class IncreaseStarRatingStatisticService implements
-    IncreaseStarRatingStatisticUseCase {
-
-    private final StarRatingStatisticDomainService starRatingStatisticDomainService;
-
-    @Override
-    public void invoke(
-        Long perfumeId,
-        int score
-    ) {
-        starRatingStatisticDomainService.increaseVotes(
-            perfumeId,
-            score
-        );
-
-    }
-
-}
