+++ src/main/java/com/pikachu/purple/application/statistic/service/starratingstatistic/GetStarRatingStatisticsService.java
@@ -1,7 +1,7 @@
-package com.pikachu.purple.application.statistic.service.application.starratingstatistic;
+package com.pikachu.purple.application.statistic.service.starratingstatistic;
 
 import com.pikachu.purple.application.statistic.port.in.starratingstatistic.GetStarRatingStatisticsUseCase;
-import com.pikachu.purple.application.statistic.service.domain.StarRatingStatisticDomainService;
+import com.pikachu.purple.application.statistic.port.out.StarRatingStatisticRepository;
 import com.pikachu.purple.domain.statistic.StarRatingStatistic;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -11,26 +11,26 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 class GetStarRatingStatisticsService implements GetStarRatingStatisticsUseCase {
 
-    private final StarRatingStatisticDomainService starRatingStatisticDomainService;
+    private final StarRatingStatisticRepository starRatingStatisticRepository;
 
     @Override
-    public Result invoke() {
-        List<StarRatingStatistic> starRatingStatistics = starRatingStatisticDomainService.findAll();
+    public Result findAll() {
+        List<StarRatingStatistic> starRatingStatistics = starRatingStatisticRepository.findAll();
 
         return new Result(starRatingStatistics);
     }
 
     @Override
-    public Result invoke(Long perfumeId) {
-        List<StarRatingStatistic> starRatingStatistics = starRatingStatisticDomainService
+    public Result findAll(Long perfumeId) {
+        List<StarRatingStatistic> starRatingStatistics = starRatingStatisticRepository
             .findAll(perfumeId);
 
         return new Result(starRatingStatistics);
     }
 
     @Override
-    public Result invoke(List<Long> perfumeIds) {
-        List<StarRatingStatistic> starRatingStatistics = starRatingStatisticDomainService
+    public Result findAll(List<Long> perfumeIds) {
+        List<StarRatingStatistic> starRatingStatistics = starRatingStatisticRepository
             .findAll(perfumeIds);
 
         return new Result(starRatingStatistics);
