+++ src/main/java/com/pikachu/purple/application/statistic/service/starratingstatistic/IncreaseStarRatingStatisticsService.java
@@ -1,7 +1,7 @@
-package com.pikachu.purple.application.statistic.service.application.starratingstatistic;
+package com.pikachu.purple.application.statistic.service.starratingstatistic;
 
+import com.pikachu.purple.application.statistic.port.in.starratingstatistic.IncreaseStarRatingStatisticUseCase;
 import com.pikachu.purple.application.statistic.port.in.starratingstatistic.IncreaseStarRatingStatisticsUseCase;
-import com.pikachu.purple.application.statistic.service.domain.StarRatingStatisticDomainService;
 import com.pikachu.purple.bootstrap.onboarding.vo.StarRatingVO;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -12,12 +12,12 @@ import org.springframework.stereotype.Service;
 class IncreaseStarRatingStatisticsService implements
     IncreaseStarRatingStatisticsUseCase {
 
-    private final StarRatingStatisticDomainService starRatingStatisticDomainService;
+    private final IncreaseStarRatingStatisticUseCase increaseStarRatingStatisticUseCase;
 
     @Override
     public void invoke(List<StarRatingVO> starRatingVOs) {
         for (StarRatingVO starRatingVO : starRatingVOs) {
-            starRatingStatisticDomainService.increaseVotes(
+            increaseStarRatingStatisticUseCase.invoke(
                 starRatingVO.perfumeId(),
                 starRatingVO.score()
             );
